+++ src/main/java/dev/wannabe/reciperepository/service/RecipeService.java
@@ -28,9 +28,15 @@ public class RecipeService {
         return recipeRepository.save(recipe);
     }
 
-    public void deleteById(Long id) {
-        recipeRepository.deleteById(id);
+    public long deleteById(Long id) {
+        boolean exists = recipeRepository.existsById(id);
+        if (exists) {
+            recipeRepository.deleteById(id);
+            return id;
+        }
+        return -1;
     }
+
 
 
 }
